Some function split, no MPI
Tried to implement MPI, but could not manage to find a good way to split the domain, will try again later. Started splitting functions into separate .cpp and .h-files and adding the to the makefile, successfully compiling all cases on macOS. Need to verify that it will work on the Linux environment too.